requirements.txt
# Telegram Bot Framework
python-telegram-bot==20.7

# Environment Variables Management
python-dotenv==1.0.0

# Database ORM
sqlalchemy==2.0.23

# Optional: For better async support
aiofiles==23.2.1